pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trisolaris"
version = "0.1.0"
description = "TRISOLARIS Framework for Evolutionary Algorithm Development"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pytest",
    "python-magic",
    "typing-extensions",
]

[tool.setuptools]
packages = ["trisolaris"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 